--- application-common/src/main/java/com/seo/dto/report/ReportsDto.java
@@ -1,10 +1,0 @@
-package com.seo.dto.report;
-
-import lombok.Data;
-
-
-@Data
-public class ReportsDto {
-    private ReportDto report;
-    // TODO: Add other report ids.
-}
