+++ application-config/src/main/java/com/seo/config/AppConfig.java
@@ -2,10 +2,10 @@ package com.seo.config;
 
 import com.seo.crawler.CrawlClient;
 import com.seo.crawler.impl.CrawlClientImpl;
-import com.seo.crawler.impl.ContentParserImpl;
-import com.seo.crawler.impl.LinkParser;
-import com.seo.crawler.Parser;
-import com.seo.crawler.ContentParser;
+import com.seo.parser.impl.ContentParserImpl;
+import com.seo.parser.impl.LinkParser;
+import com.seo.parser.Parser;
+import com.seo.parser.ContentParser;
 import com.seo.model.CrawlData;
 import com.seo.crawler.impl.CrawlTask;
 import org.springframework.beans.factory.config.BeanDefinition;
