+++ application-crawler/src/main/java/com/seo/model/CrawlResponse.java
@@ -1,4 +1,4 @@
-package com.seo.crawl;
+package com.seo.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
