+++ application-crawler/src/main/java/com/seo/parser/ContentParser.java
@@ -1,4 +1,4 @@
-package com.seo.crawler;
+package com.seo.parser;
 
 import com.seo.model.Content;
