+++ application-crawler/src/main/java/com/seo/service/CrawlerService.java
@@ -1,10 +1,10 @@
 package com.seo.service;
 
 import com.seo.ContentEntity;
-import com.seo.crawl.BasicCrawlStatus;
-import com.seo.crawl.CrawlRequest;
-import com.seo.crawl.CrawlResponse;
-import com.seo.crawl.CrawlStatus;
+import com.seo.model.BasicCrawlStatus;
+import com.seo.model.CrawlRequest;
+import com.seo.model.CrawlResponse;
+import com.seo.model.CrawlStatus;
 import com.seo.crawler.CrawlCompleteListener;
 import com.seo.crawler.impl.CrawlTask;
 import com.seo.exception.TaskNotFoundException;
