+++ application-persistence/src/main/java/com/seo/repository/report/ReportDocumentRepository.java
@@ -6,7 +6,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface ReportDocumentRepository extends MongoRepository<BaseReportDocument, String> {
     Page<BaseReportDocument> findAllByWebsiteProjectId(String websiteProjectId, Pageable pageable);
+    List<BaseReportDocument> findAllByWebsiteProjectId(String websiteProjectId);
 }
